[notify] xkbswitch: removed (upstream transition)
`xkbswitch` was renamed to `rawkb` and moved from Zeppe-Lin Org.

This decision stems from a shift in focus and interest.

What does this means for you? If you are currently using the `xkbswitch`
package from Zeppe-Lin, it will no longer receive updates or be
available for installation from our official pkgsrc repositories.

For those who appreciate and wish to continue using this, it has been
renamed to `rawkb` and has found a new home at:
<https://github.com/rawmkit/rawkb>.  This new repository will be where
any future development or maintenance of this software will occur if
anyone is interested in.

I encourage interested users to migrate to the `rawkb` project directly,
and (optionally) send a patch with new *package source* to
<zeppe-lin-dev@freelists.org> or open PR at
<https://github.com/zeppe-lin/pkgsrc-xorg>.